+++ backend/spec/features/admin/orders/shipments_spec.rb
@@ -41,14 +41,13 @@ describe "Shipments" do
       targetted_select2 'LA', from: '#s2id_item_stock_location'
       click_icon :ok
       wait_for_ajax
-      sleep(1)
-      page.should have_selector("#shipment_#{order.shipments.last.id}")
+      page.should have_selector("#shipment_#{order.shipments.first.id}")
 
       within_row(2) { click_icon 'resize-horizontal' }
       targetted_select2 "LA(#{order.reload.shipments.last.number})", from: '#s2id_item_stock_location'
       click_icon :ok
       wait_for_ajax
-      page.should have_selector("#shipment_#{order.shipments.last.id}")
+      page.should have_selector("#shipment_#{order.reload.shipments.last.id}")
     end
   end
 end