+++ core/lib/spree/testing_support/factories/product_factory.rb
@@ -7,34 +7,24 @@ FactoryGirl.define do
     sku 'ABC'
     available_on { 1.year.ago }
     deleted_at nil
+    shipping_category { |r| Spree::ShippingCategory.first || r.association(:shipping_category) }
 
     # ensure stock item will be created for this products master
     before(:create) { create(:stock_location) if Spree::StockLocation.count == 0 }
 
+    factory :custom_product do
+      name 'Custom Product'
+      price 17.99
+
+      tax_category { |r| Spree::TaxCategory.first || r.association(:tax_category) }
+    end
+
     factory :product do
       tax_category { |r| Spree::TaxCategory.first || r.association(:tax_category) }
-      shipping_category { |r| Spree::ShippingCategory.first || r.association(:shipping_category) }
 
       factory :product_with_option_types do
         after(:create) { |product| create(:product_option_type, product: product) }
       end
     end
-  end
-
-  factory :custom_product, class: Spree::Product do
-    name 'Custom Product'
-    description { generate(:random_description) }
-    price 17.99
-    sku 'ABC'
-    available_on { 1.year.ago }
-    deleted_at nil
-
-    tax_category { |r| Spree::TaxCategory.first || r.association(:tax_category) }
-    shipping_category { |r| Spree::ShippingCategory.first || r.association(:shipping_category) }
-
-    # association :taxons
-
-    # ensure stock item will be created for this products master
-    before(:create) { create(:stock_location) if Spree::StockLocation.count == 0 }
   end
 end