+++ api/CHANGELOG.md
@@ -7,3 +7,7 @@
 * Checkouts API's update action will now correctly process line item attributes (either `line_items` or `line_item_attributes`)
 
     * Ryan Bigg*
+
+* Checkouts API now correctly processes incoming payment data during the payment step.
+
+    *Ryan Bigg*